Updating HoldController test to include HTTP POST on /protected/cloneable endpoint
